requirements.txt
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
scipy>=0.15.1
Pillow==8.1.1
numpy>=1.9.2
scikit-learn>=0.18
pandas>=0.19
